requirements.txt
django==5.0.7
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
celery==5.3.6
redis==5.0.7
google-generativeai==0.7.2
python-dotenv==1.0.1
django-cors-headers
pyJWT
pytz
asgiref
gunicorn
sqlparse